+++ src/backend/resolvers/mutations/sitemap/upsertUsersSitemaps/upsertUserSitemap/getCombinedSitemaps/index.js
@@ -1,10 +1,10 @@
 /* eslint-disable no-loop-func */
 const _ = require('lodash')
 
-module.exports = sitemaps => {
+module.exports = (sitemaps) => {
   const combinedSitemaps = {}
   for (const sitemap of sitemaps) {
-    Object.keys(sitemap).forEach(nodeType => {
+    Object.keys(sitemap).forEach((nodeType) => {
       const nodes = sitemap[nodeType]
 
       if (!combinedSitemaps[nodeType]) {
@@ -12,17 +12,14 @@ module.exports = sitemaps => {
       } else {
         combinedSitemaps[nodeType] = combinedSitemaps[nodeType].concat(nodes)
 
-        combinedSitemaps[nodeType] = _.uniqBy(
-          combinedSitemaps[nodeType],
-          '_id'
-        )
+        combinedSitemaps[nodeType] = _.uniqBy(combinedSitemaps[nodeType], '_id')
       }
-      combinedSitemaps[nodeType] = _.orderBy(
-        combinedSitemaps[nodeType],
-        ['parentId', 'order']
-      )
+      combinedSitemaps[nodeType] = _.orderBy(combinedSitemaps[nodeType], [
+        'parentId',
+        'order',
+      ])
     })
   }
 
-  return combinedSitemaps;
+  return combinedSitemaps
 }