+++ src/backend/resolvers/mutations/sitemap/upsertUsersSitemaps/upsertUserSitemap/getCombinedSitemaps/tests/mocks/input/disjoint-sitemaps-with-missing-node-types.js
@@ -1,0 +1,15 @@
+const {
+  PAYER_MANAGEMENT_SUMMARY_PAGE,
+  PAYER_MANAGEMENT_DASHBOARD,
+} = require('../../../../../../../shared/mocks/nodes')
+
+module.exports = [
+  {
+    tools: [],
+    pages: [PAYER_MANAGEMENT_SUMMARY_PAGE],
+  },
+  {
+    dashboards: [PAYER_MANAGEMENT_DASHBOARD],
+    cards: [],
+  },
+]