+++ src/backend/resolvers/mutations/sitemap/upsertUsersSitemaps/upsertUserSitemap/getCombinedSitemaps/tests/mocks/input/two-disjoint-sitemaps.js
@@ -1,0 +1,3 @@
+const { ROLE_A, ROLE_B } = require('../../../../mocksInput/roles')
+
+module.exports = [ROLE_A.sitemap, ROLE_B.sitemap]