+++ src/backend/resolvers/mutations/sitemap/upsertUsersSitemaps/upsertUserSitemap/getCombinedSitemaps/tests/mocks/input/two-joint-sitemaps-1.js
@@ -1,0 +1,3 @@
+const { ROLE_A, ROLE_C } = require('../../../../mocksInput/roles')
+
+module.exports = [ROLE_A.sitemap, ROLE_C.sitemap]