+++ src/backend/resolvers/mutations/sitemap/upsertUsersSitemaps/upsertUserSitemap/getCombinedSitemaps/tests/mocks/output/disjoint-sitemaps-with-missing-node-types.js.js
@@ -1,0 +1,11 @@
+const {
+  PAYER_MANAGEMENT_SUMMARY_PAGE,
+  PAYER_MANAGEMENT_DASHBOARD,
+} = require('../../../../../../../shared/mocks/nodes')
+
+module.exports = {
+  tools: [],
+  dashboards: [PAYER_MANAGEMENT_DASHBOARD],
+  pages: [PAYER_MANAGEMENT_SUMMARY_PAGE],
+  cards: [],
+}